+++ event-store-management/event-store-management-core/src/test/java/uk/gov/justice/services/eventstore/management/catchup/state/CatchupStateManagerTest.java
@@ -1,11 +1,13 @@
 package uk.gov.justice.services.eventstore.management.catchup.state;
 
+import static org.hamcrest.CoreMatchers.hasItems;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import uk.gov.justice.services.common.util.UtcClock;
+import uk.gov.justice.services.eventstore.management.catchup.process.CatchupFor;
 import uk.gov.justice.services.eventstore.management.catchup.process.CatchupInProgress;
 import uk.gov.justice.services.eventstore.management.commands.EventCatchupCommand;
 import uk.gov.justice.services.eventstore.management.commands.IndexerCatchupCommand;
@@ -13,7 +15,6 @@ import uk.gov.justice.services.eventstore.management.commands.IndexerCatchupComm
 import java.time.ZonedDateTime;
 import java.util.List;
 
-import org.hamcrest.CoreMatchers;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
@@ -31,7 +32,7 @@ public class CatchupStateManagerTest {
         final EventCatchupCommand eventCatchupCommand = new EventCatchupCommand();
 
         final CatchupInProgress indexCatchupInProgress = mock(CatchupInProgress.class);
-        when(indexCatchupInProgress.getSubscriptionName()).thenReturn("different_catchup");
+        when(indexCatchupInProgress.getCatchupFor()).thenReturn(mock(CatchupFor.class));
 
         catchupStateManager.addCatchupInProgress(indexCatchupInProgress, new IndexerCatchupCommand());
 
@@ -39,35 +40,39 @@ public class CatchupStateManagerTest {
 
         final ZonedDateTime startedAt = new UtcClock().now();
 
-        final CatchupInProgress catchupInProgress_1 = new CatchupInProgress("subscription_1", startedAt);
-        final CatchupInProgress catchupInProgress_2 = new CatchupInProgress("subscription_2", startedAt.plusMinutes(1));
-        final CatchupInProgress catchupInProgress_3 = new CatchupInProgress("subscription_3", startedAt.plusMinutes(2));
+        final CatchupFor catchupFor_1 = new CatchupFor("subscription_1", "component_1");
+        final CatchupFor catchupFor_2 = new CatchupFor("subscription_2", "component_1");
+        final CatchupFor catchupFor_3 = new CatchupFor("subscription_3", "component_2");
+
+        final CatchupInProgress catchupInProgress_1 = new CatchupInProgress(catchupFor_1, startedAt);
+        final CatchupInProgress catchupInProgress_2 = new CatchupInProgress(catchupFor_2, startedAt.plusMinutes(1));
+        final CatchupInProgress catchupInProgress_3 = new CatchupInProgress(catchupFor_3, startedAt.plusMinutes(2));
 
         catchupStateManager.addCatchupInProgress(catchupInProgress_1, eventCatchupCommand);
         catchupStateManager.addCatchupInProgress(catchupInProgress_2, eventCatchupCommand);
         catchupStateManager.addCatchupInProgress(catchupInProgress_3, eventCatchupCommand);
 
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_1.getSubscriptionName(), eventCatchupCommand), is(true));
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_2.getSubscriptionName(), eventCatchupCommand), is(true));
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_3.getSubscriptionName(), eventCatchupCommand), is(true));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_1.getCatchupFor(), eventCatchupCommand), is(true));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_2.getCatchupFor(), eventCatchupCommand), is(true));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_3.getCatchupFor(), eventCatchupCommand), is(true));
 
         final List<CatchupInProgress> allCatchupsInProgress = catchupStateManager.getAllCatchupsInProgress(eventCatchupCommand);
 
         assertThat(allCatchupsInProgress.size(), is(3));
-        assertThat(allCatchupsInProgress, CoreMatchers.hasItems(catchupInProgress_1, catchupInProgress_2, catchupInProgress_3));
+        assertThat(allCatchupsInProgress, hasItems(catchupInProgress_1, catchupInProgress_2, catchupInProgress_3));
 
-        final CatchupInProgress removedCatchupInProgress = catchupStateManager.removeCatchupInProgress(catchupInProgress_2.getSubscriptionName(), eventCatchupCommand);
+        final CatchupInProgress removedCatchupInProgress = catchupStateManager.removeCatchupInProgress(catchupInProgress_2.getCatchupFor(), eventCatchupCommand);
 
         assertThat(removedCatchupInProgress, is(catchupInProgress_2));
 
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_1.getSubscriptionName(), eventCatchupCommand), is(true));
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_2.getSubscriptionName(), eventCatchupCommand), is(false));
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_3.getSubscriptionName(), eventCatchupCommand), is(true));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_1.getCatchupFor(), eventCatchupCommand), is(true));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_2.getCatchupFor(), eventCatchupCommand), is(false));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_3.getCatchupFor(), eventCatchupCommand), is(true));
 
         final List<CatchupInProgress> currentCatchupsInProgress = catchupStateManager.getAllCatchupsInProgress(eventCatchupCommand);
 
         assertThat(currentCatchupsInProgress.size(), is(2));
-        assertThat(currentCatchupsInProgress, CoreMatchers.hasItems(catchupInProgress_1, catchupInProgress_3));
+        assertThat(currentCatchupsInProgress, hasItems(catchupInProgress_1, catchupInProgress_3));
 
         catchupStateManager.clear(eventCatchupCommand);
 
@@ -81,7 +86,7 @@ public class CatchupStateManagerTest {
         final IndexerCatchupCommand indexerCatchupCommand = new IndexerCatchupCommand();
 
         final CatchupInProgress eventCatchupInProgress = mock(CatchupInProgress.class);
-        when(eventCatchupInProgress.getSubscriptionName()).thenReturn("different_catchup");
+        when(eventCatchupInProgress.getCatchupFor()).thenReturn(mock(CatchupFor.class));
 
         catchupStateManager.addCatchupInProgress(eventCatchupInProgress, new EventCatchupCommand());
 
@@ -89,35 +94,39 @@ public class CatchupStateManagerTest {
 
         final ZonedDateTime startedAt = new UtcClock().now();
 
-        final CatchupInProgress catchupInProgress_1 = new CatchupInProgress("subscription_1", startedAt);
-        final CatchupInProgress catchupInProgress_2 = new CatchupInProgress("subscription_2", startedAt.plusMinutes(1));
-        final CatchupInProgress catchupInProgress_3 = new CatchupInProgress("subscription_3", startedAt.plusMinutes(2));
+        final CatchupFor catchupFor_1 = new CatchupFor("subscription_1", "component_1");
+        final CatchupFor catchupFor_2 = new CatchupFor("subscription_2", "component_1");
+        final CatchupFor catchupFor_3 = new CatchupFor("subscription_3", "component_2");
+
+        final CatchupInProgress catchupInProgress_1 = new CatchupInProgress(catchupFor_1, startedAt);
+        final CatchupInProgress catchupInProgress_2 = new CatchupInProgress(catchupFor_2, startedAt.plusMinutes(1));
+        final CatchupInProgress catchupInProgress_3 = new CatchupInProgress(catchupFor_3, startedAt.plusMinutes(2));
 
         catchupStateManager.addCatchupInProgress(catchupInProgress_1, indexerCatchupCommand);
         catchupStateManager.addCatchupInProgress(catchupInProgress_2, indexerCatchupCommand);
         catchupStateManager.addCatchupInProgress(catchupInProgress_3, indexerCatchupCommand);
 
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_1.getSubscriptionName(), indexerCatchupCommand), is(true));
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_2.getSubscriptionName(), indexerCatchupCommand), is(true));
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_3.getSubscriptionName(), indexerCatchupCommand), is(true));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_1.getCatchupFor(), indexerCatchupCommand), is(true));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_2.getCatchupFor(), indexerCatchupCommand), is(true));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_3.getCatchupFor(), indexerCatchupCommand), is(true));
 
         final List<CatchupInProgress> allCatchupsInProgress = catchupStateManager.getAllCatchupsInProgress(indexerCatchupCommand);
 
         assertThat(allCatchupsInProgress.size(), is(3));
-        assertThat(allCatchupsInProgress, CoreMatchers.hasItems(catchupInProgress_1, catchupInProgress_2, catchupInProgress_3));
+        assertThat(allCatchupsInProgress, hasItems(catchupInProgress_1, catchupInProgress_2, catchupInProgress_3));
 
-        final CatchupInProgress removedCatchupInProgress = catchupStateManager.removeCatchupInProgress(catchupInProgress_2.getSubscriptionName(), indexerCatchupCommand);
+        final CatchupInProgress removedCatchupInProgress = catchupStateManager.removeCatchupInProgress(catchupInProgress_2.getCatchupFor(), indexerCatchupCommand);
 
         assertThat(removedCatchupInProgress, is(catchupInProgress_2));
 
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_1.getSubscriptionName(), indexerCatchupCommand), is(true));
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_2.getSubscriptionName(), indexerCatchupCommand), is(false));
-        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_3.getSubscriptionName(), indexerCatchupCommand), is(true));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_1.getCatchupFor(), indexerCatchupCommand), is(true));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_2.getCatchupFor(), indexerCatchupCommand), is(false));
+        assertThat(catchupStateManager.isCatchupInProgress(catchupInProgress_3.getCatchupFor(), indexerCatchupCommand), is(true));
 
         final List<CatchupInProgress> currentCatchupsInProgress = catchupStateManager.getAllCatchupsInProgress(indexerCatchupCommand);
 
         assertThat(currentCatchupsInProgress.size(), is(2));
-        assertThat(currentCatchupsInProgress, CoreMatchers.hasItems(catchupInProgress_1, catchupInProgress_3));
+        assertThat(currentCatchupsInProgress, hasItems(catchupInProgress_1, catchupInProgress_3));
 
         catchupStateManager.clear(indexerCatchupCommand);